+++ controllers/errors.js
@@ -1,0 +1,5 @@
+exports.get404 = (req, res, next) => {
+    res.status(404).render('404', {
+        pageTitle: '404 - My Shop!',
+    });
+}
