+++ routes/admin.js
@@ -1,24 +1,14 @@
 const express = require('express');
 const path = require('path');
 
-const rootDir = require('../utils/path');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-const products = [];
-
 // (GET)/admin/new-product => New product form
-router.get('/new-product', (req, res, next) => {
-    res.render('new-product', {
-        pageTitle: 'Create a new product - My Shop!',
-    });
-});
+router.get('/new-product', productsController.getAddProduct);
 
 // (POST)/admin/new-product => Creation of new product and redirect to /
-router.post('/new-product', (req, res, next) => {
-    products.push({title: req.body.title});
-    res.redirect('/');
-});
+router.post('/new-product', productsController.postAddProduct);
 
-exports.router = router;
-exports.products = products;
+module.exports = router;
